requirements.txt
fastapi
uvicorn
streamlit
transformers
torch
Pillow
requests